Added multiple tokenizers implementing new ITokenizer interface (GetOpt = previous, Windows, Hybrid)
