+++ src/CommandLine/Core/ITokenizer.cs
@@ -1,0 +1,26 @@
+﻿// Copyright 2005-2015 Giacomo Stelluti Scala & Contributors. All rights reserved. See doc/License.md in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using CommandLine.Infrastructure;
+using CSharpx;
+using RailwaySharp.ErrorHandling;
+
+namespace CommandLine.Core
+{
+    internal interface ITokenizer
+    {
+        Result<IEnumerable<Token>, Error> Tokenize(
+            IEnumerable<string> arguments,
+            Func<string, bool> nameLookup);
+
+        Result<IEnumerable<Token>, Error> Preprocess(
+            IEnumerable<string> arguments,
+            Func<IEnumerable<string>, Result<IEnumerable<Token>, Error>> tokenizer);
+
+        Result<IEnumerable<Token>, Error> ExplodeOptionList(
+            Result<IEnumerable<Token>, Error> tokenizerResult,
+            Func<string, Maybe<char>> optionSequenceWithSeparatorLookup);
+    }
+}
